Fix ubuntu build, which fails on comparison between different signedness

+++ applications/GeoMechanicsApplication/tests/cpp_tests/stub_time_loop_executor.cpp
@@ -15,7 +15,7 @@
 
 namespace Kratos {
 
-StubTimeLoopExecutor::StubTimeLoopExecutor(int NumberOfExpectedProcesses) :
+StubTimeLoopExecutor::StubTimeLoopExecutor(size_t NumberOfExpectedProcesses) :
         mNumberOfExpectedProcesses{NumberOfExpectedProcesses}
 {
 }
